+++ src/main/resources/db/changelog/dictionaries/scenario_group.sql
@@ -3,7 +3,7 @@ CREATE TABLE IF NOT EXISTS bing.scenario_group
     id           UUID            PRIMARY KEY,
     name         VARCHAR(255)    NOT NULL,
     object_id    UUID            NOT NULL,
-    danger_koef  INTEGER         NOT NULL,
+    danger_koef  FLOAT         NOT NULL,
 
     CONSTRAINT fk_object
         FOREIGN KEY (object_id)
